Copy files to destination without metadata
This is to support mounts to SAMBA shares or NTFS partitions. Those
filesystems do not support file permission bitflags or other POSIX
metadata. `shutil.move`, `shutil.copy2` and `shutil.copy` all copy the
file *and then* attempt to do `chmod` or copy stats over.

This change literally only copies the files. Permission flags are
untouched. This shouldn't be a problem in theory since the destination
will either not need that information *or* have inheritable permissions
already properly set.

Without this change, previous behavior resulted in copies properly
happening but nzbget would show them as post-processing failures
because the `copystat` piece was failing, if the target directory was
a SAMBA share.