feat: Add support to mark specifiers as deprecated
Across the formats, different specifiers are either buggy or undocumented.
To provide a safe way for users to migrate, a deprecation notice has been added.

Mapping tuples now can have a third field with a deprecation message that is
shown to the user as Warning, when the specifier is replaced.

See: #52